rename and restructure condition hover component
