+++ src/components/card.tsx/card.tsx
@@ -6,7 +6,7 @@ import * as Collapsible from '@radix-ui/react-collapsible'
 import * as Separator from '@radix-ui/react-separator'
 import { useEffect, useState } from 'react'
 import CardHeader from '../card-header'
-import { ParsedDescription } from '../parsed-description.tsx/parsed-description'
+import { ParsedDescription } from '../parsed-description/parsed-description'
 import SourceDisplay from '../source-display'
 import Traits from '../traits-display'
 import styles from './card.module.css'
