+++ src/components/equipment-display.tsx
@@ -2,7 +2,7 @@
 
 import { Equipment } from '@/models/equipment'
 import { useEffect, useState } from 'react'
-import CardDisplayList from './card-display-list.tsx/card-display-list'
+import CardDisplayList from './card-display-list/card-display-list'
 import EquipmentCard from './equipment-card'
 import SelectableGrid from './selectable-grid/selectable-grid'
